@@ -1,44 +1,41 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:effektio/common/store/themes/AppTheme.dart';
 import 'package:effektio/common/store/themes/SeperatedThemes.dart';
 import 'package:effektio/controllers/chat_list_controller.dart';
 import 'package:effektio/controllers/chat_room_controller.dart';
-import 'package:effektio/controllers/network_controller.dart';
 import 'package:effektio/controllers/receipt_controller.dart';
 import 'package:effektio/l10n/l10n.dart';
 import 'package:effektio/screens/HomeScreens/chat/Overview.dart';
 // import 'package:effektio/screens/HomeScreens/Notification.dart';
 import 'package:effektio/screens/HomeScreens/faq/Overview.dart';
 import 'package:effektio/screens/HomeScreens/news/News.dart';
-import 'package:effektio/screens/HomeScreens/todo/AddToDo.dart';
 import 'package:effektio/screens/HomeScreens/todo/ToDo.dart';
 import 'package:effektio/screens/OnboardingScreens/LogIn.dart';
 import 'package:effektio/screens/OnboardingScreens/Signup.dart';
 import 'package:effektio/screens/SideMenuScreens/Gallery.dart';
 import 'package:effektio/screens/UserScreens/SocialProfile.dart';
 import 'package:effektio/widgets/AppCommon.dart';
+// import 'package:effektio/widgets/AppCommon.dart';
 import 'package:effektio/widgets/CrossSigning.dart';
 import 'package:effektio/widgets/MaterialIndicator.dart';
 import 'package:effektio/widgets/SideMenu.dart';
 import 'package:effektio_flutter_sdk/effektio_flutter_sdk.dart'
     show Client, EffektioSdk;
 import 'package:effektio_flutter_sdk/effektio_flutter_sdk_ffi.dart'
-    show FfiBufferUint8, SyncState;
+    show CreateGroupSettings, FfiBufferUint8, SyncState;
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:flutter_icons_null_safety/flutter_icons_null_safety.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_mentions/flutter_mentions.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:get/get.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:flutter_icons_null_safety/flutter_icons_null_safety.dart';
 import 'package:overlay_support/overlay_support.dart';
 import 'package:themed/themed.dart';
-import 'package:window_size/window_size.dart';
 
 void main() async {
   await startApp();
@@ -46,10 +43,6 @@ void main() async {
 
 Future<void> startApp() async {
   WidgetsFlutterBinding.ensureInitialized();
-  bool isDesktop = Platform.isWindows || Platform.isMacOS || Platform.isLinux;
-  if (isDesktop) {
-    setWindowTitle('Effektio');
-  }
   GoogleFonts.config.allowRuntimeFetching = false;
   LicenseRegistry.addLicense(() async* {
     final license = await rootBundle.loadString('google_fonts/LICENSE.txt');
@@ -86,8 +79,6 @@ class Effektio extends StatelessWidget {
               '/profile': (BuildContext context) => const SocialProfileScreen(),
               '/signup': (BuildContext context) => const SignupScreen(),
               '/gallery': (BuildContext context) => const GalleryScreen(),
-              '/todo': (BuildContext context) => const ToDoScreen(),
-              '/addTodo': (BuildContext context) => const AddToDoScreen(),
             },
           ),
         ),
@@ -111,11 +102,11 @@ class _EffektioHomeState extends State<EffektioHome>
   late TabController tabController;
   String? displayName;
   Future<FfiBufferUint8>? displayAvatar;
-  final networkController = Get.put(NetworkController());
 
   @override
   void initState() {
     super.initState();
+
     client = makeClient();
     tabController = TabController(length: 4, vsync: this);
     tabController.addListener(() {
@@ -144,6 +135,20 @@ class _EffektioHomeState extends State<EffektioHome>
     syncState = client.startSync();
     //Start listening for cross signing events
     if (!client.isGuest()) {
+      await client.groups().then(
+        (groups) async {
+          if (groups.toList().isEmpty) {
+            // Create default effektio group when client synced.
+            CreateGroupSettings settings = sdk.newGroupSettings(
+              '${simplifyUserId(client.userId().toString())} Team',
+            );
+            settings.alias(UniqueKey().toString());
+            settings.visibility('Public');
+            settings.addInvitee('@sisko:matrix.org');
+            await client.createEffektioGroup(settings);
+          }
+        },
+      );
       await client.getUserProfile().then((value) {
         if (mounted) {
           setState(() {
@@ -275,52 +280,49 @@ class _EffektioHomeState extends State<EffektioHome>
       length: 4,
       key: const Key('bottom-bar'),
       child: SafeArea(
-        child: Stack(
-          children: [
-            Scaffold(
-              appBar: buildAppBar(),
-              body: TabBarView(
-                controller: tabController,
-                children: [
-                  NewsScreen(
-                    client: client,
-                    displayName: displayName,
-                    displayAvatar: displayAvatar,
-                  ),
-                  FaqOverviewScreen(client: client),
-                  const ToDoScreen(),
-                  ChatOverview(client: client),
-                ],
-              ),
-              drawer: SideDrawer(
-                isGuest: client.isGuest(),
-                userId: client.userId().toString(),
+        child: Scaffold(
+          appBar: buildAppBar(),
+          body: TabBarView(
+            controller: tabController,
+            children: [
+              NewsScreen(
+                client: client,
                 displayName: displayName,
                 displayAvatar: displayAvatar,
               ),
-              bottomNavigationBar: TabBar(
-                labelColor: AppCommonTheme.primaryColor,
-                unselectedLabelColor: AppCommonTheme.svgIconColor,
-                controller: tabController,
-                indicator: const MaterialIndicator(
-                  height: 5,
-                  bottomLeftRadius: 8,
-                  bottomRightRadius: 8,
-                  topLeftRadius: 0,
-                  topRightRadius: 0,
-                  horizontalPadding: 12,
-                  tabPosition: TabPosition.top,
-                  color: AppCommonTheme.primaryColor,
-                ),
-                tabs: [
-                  buildNewsFeedTab(),
-                  buildPinsTab(),
-                  buildTasksTab(),
-                  buildChatTab(),
-                ],
-              ),
+              FaqOverviewScreen(client: client),
+              ToDoScreen(client: client),
+              ChatOverview(client: client),
+            ],
+          ),
+          drawer: SideDrawer(
+            isGuest: client.isGuest(),
+            userId: client.userId().toString(),
+            displayName: displayName,
+            displayAvatar: displayAvatar,
+          ),
+          bottomNavigationBar: TabBar(
+            labelColor: AppCommonTheme.primaryColor,
+            unselectedLabelColor: AppCommonTheme.svgIconColor,
+            controller: tabController,
+            indicatorSize: TabBarIndicatorSize.tab,
+            indicatorPadding: const EdgeInsets.symmetric(horizontal: 12),
+            indicator: const MaterialIndicator(
+              height: 5,
+              bottomLeftRadius: 8,
+              bottomRightRadius: 8,
+              topLeftRadius: 0,
+              topRightRadius: 0,
+              tabPosition: TabPosition.top,
+              color: AppCommonTheme.primaryColor,
             ),
-          ],
+            tabs: [
+              buildNewsFeedTab(),
+              buildPinsTab(),
+              buildTasksTab(),
+              buildChatTab(),
+            ],
+          ),
         ),
       ),
     );
@@ -355,5 +357,4 @@ class _EffektioHomeState extends State<EffektioHome>
       },
     );
   }
-
 }
