@@ -17,8 +17,8 @@ class _InviteUserDialogState extends State<InviteUserDialog> {
         children: [
           Container(
             padding: const EdgeInsets.symmetric(
-              horizontal: 16.0,
-              vertical: 24.0,
+              horizontal: 16,
+              vertical: 24,
             ),
             child: Center(
               child: Column(
@@ -28,35 +28,49 @@ class _InviteUserDialogState extends State<InviteUserDialog> {
                     'Invite Friends',
                     style: ToDoTheme.titleTextStyle.copyWith(fontSize: 18),
                   ),
-                  const SizedBox(
-                    height: 4,
-                  ),
+                  const SizedBox(height: 4),
                   Text(
                     'You can invite your friends to ToDo today via',
                     textAlign: TextAlign.center,
-                    style: ToDoTheme.subtitleTextStyle.copyWith(
+                    style: ToDoTheme.descriptionTextStyle.copyWith(
                       color: ToDoTheme.calendarColor,
-                      fontSize: 15,),
+                      fontSize: 15,
+                    ),
                   ),
-                  const SizedBox(
-                    height: 16,
-                  ),
+                  const SizedBox(height: 16),
                   Column(
                     children: [
-                      buildDivider(),
-                      Text('Whatsapp', style : ToDoTheme.titleTextStyle.copyWith(fontSize: 16)),
-                      buildDivider(),
-                      Text('Email', style : ToDoTheme.titleTextStyle.copyWith(fontSize: 16)),
-                      buildDivider(),
-                      Text('SMS', style : ToDoTheme.titleTextStyle.copyWith(fontSize: 16)),
-                      buildDivider(),
-                      Text('Invitation Link', style : ToDoTheme.titleTextStyle.copyWith(fontSize: 16)),
-                      buildDivider(),
+                      const _Divider(),
+                      Text(
+                        'Whatsapp',
+                        style: ToDoTheme.titleTextStyle.copyWith(fontSize: 16),
+                      ),
+                      const _Divider(),
+                      Text(
+                        'Email',
+                        style: ToDoTheme.titleTextStyle.copyWith(fontSize: 16),
+                      ),
+                      const _Divider(),
+                      Text(
+                        'SMS',
+                        style: ToDoTheme.titleTextStyle.copyWith(fontSize: 16),
+                      ),
+                      const _Divider(),
+                      Text(
+                        'Invitation Link',
+                        style: ToDoTheme.titleTextStyle.copyWith(fontSize: 16),
+                      ),
+                      const _Divider(),
                       GestureDetector(
-                        onTap: (){
+                        onTap: () {
                           Navigator.pop(context);
                         },
-                        child: Text('Cancel', style : ToDoTheme.titleTextStyle.copyWith(fontSize: 16, color: Colors.red)),)
+                        child: Text(
+                          'Cancel',
+                          style: ToDoTheme.titleTextStyle
+                              .copyWith(fontSize: 16, color: Colors.red),
+                        ),
+                      )
                     ],
                   ),
                 ],
@@ -67,10 +81,15 @@ class _InviteUserDialogState extends State<InviteUserDialog> {
       ),
     );
   }
+}
 
-  Widget buildDivider() {
+class _Divider extends StatelessWidget {
+  const _Divider();
+
+  @override
+  Widget build(BuildContext context) {
     return const Padding(
-      padding: EdgeInsets.symmetric(vertical: 12.0),
+      padding: EdgeInsets.symmetric(vertical: 12),
       child: Divider(
         height: 2,
         indent: 0,
