@@ -1,295 +1,386 @@
 import 'package:effektio/common/store/themes/SeperatedThemes.dart';
 import 'package:effektio/controllers/todo_controller.dart';
-import 'package:effektio/screens/HomeScreens/todo/screens/CommentsScreen.dart';
+import 'package:effektio/models/Team.dart';
+import 'package:effektio/models/ToDoList.dart';
+import 'package:effektio/models/ToDoTask.dart';
 import 'package:effektio/widgets/AddTaskDialog.dart';
 import 'package:effektio/widgets/ExpandableText.dart';
-import 'package:effektio/widgets/ToDoTaskItem.dart';
+import 'package:effektio/widgets/ToDoTaskView.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_icons_null_safety/flutter_icons_null_safety.dart';
+import 'package:flutter_svg/flutter_svg.dart';
 import 'package:get/get.dart';
 
-class ToDoListView extends StatefulWidget {
-  final String title;
-  final String subtitle;
-  final int idx;
-
-  const ToDoListView({
-    Key? key,
-    required this.idx,
-    required this.title,
-    required this.subtitle,
-  }) : super(key: key);
-
-  @override
-  State<ToDoListView> createState() => _ToDoListViewState();
-}
-
-class _ToDoListViewState extends State<ToDoListView> {
-  ToDoController todoController = ToDoController.instance;
-  late List<ToDoTaskItem> completedTasks;
-  late List<ToDoTaskItem> pendingTasks;
-
-  @override
-  void initState() {
-    super.initState();
-    completedTasks =
-        todoController.tasks[widget.idx].where((w) => w.isCompleted).toList();
-    pendingTasks =
-        todoController.tasks[widget.idx].where((w) => !w.isCompleted).toList();
-
-    todoController.tasks[widget.idx].listen(
-      (p0) => setState(() {
-        completedTasks = todoController.tasks[widget.idx]
-            .where((w) => w.isCompleted)
-            .toList();
-        pendingTasks = todoController.tasks[widget.idx]
-            .where((w) => !w.isCompleted)
-            .toList();
-      }),
-    );
-  }
-
+class ToDoListView extends StatelessWidget {
+  const ToDoListView({Key? key, required this.controller}) : super(key: key);
+  final ToDoController controller;
   @override
   Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () {
-        todoController.toggleExpand();
-      },
-      child: Card(
-        elevation: 0,
-        margin: const EdgeInsets.all(8.0),
-        shape:
-            RoundedRectangleBorder(borderRadius: BorderRadius.circular(15.0)),
-        color: ToDoTheme.secondaryColor,
-        child: Obx(
-          () => Padding(
-            padding:
-                const EdgeInsets.symmetric(vertical: 8.0, horizontal: 15.0),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Padding(
-                  padding:
-                      const EdgeInsets.symmetric(vertical: 10, horizontal: 5),
-                  child: Row(
-                    children: [
-                      Flexible(
-                        child: Text(
-                          widget.title,
-                          style: ToDoTheme.listTitleTextStyle,
-                        ),
-                      ),
-                      const SizedBox(
-                        width: 8.0,
-                      ),
-                      Container(
-                        padding: const EdgeInsets.all(8.0),
-                        decoration: const BoxDecoration(
-                          color: AppCommonTheme.secondaryColor,
-                          borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                        ),
-                        child: const Text(
-                          'Design',
-                          style: ToDoTheme.listTagTextStyle,
-                        ),
-                      )
-                    ],
-                  ),
-                ),
-                buildSubtitle(),
-                buildDivider(),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Row(
-                      children: [
-                        const Icon(
-                          FlutterIcons.heart_evi,
-                          color: ToDoTheme.primaryTextColor,
-                        ),
-                        Text(
-                          '${todoController.likeCount}',
-                          style: ToDoTheme.calendarTextStyle,
-                        ),
-                        const SizedBox(
-                          width: 8,
-                        ),
-                        InkWell(
-                          onTap: () {
-                            Navigator.push(
-                              context,
-                              MaterialPageRoute(
-                                builder: (context) => const ToDoCommentScreen(),
-                              ),
-                            );
-                          },
-                          child: Row(
-                            children: [
-                              const Icon(
-                                FlutterIcons.comment_evi,
-                                color: ToDoTheme.primaryTextColor,
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.symmetric(horizontal: 5),
-                                child: Text(
-                                  '${todoController.messageCount}',
-                                  style: ToDoTheme.calendarTextStyle,
-                                ),
-                              ),
-                            ],
-                          ),
-                        )
-                      ],
-                    ),
-                    const Icon(
-                      FlutterIcons.bookmark_border_mdi,
-                      color: ToDoTheme.primaryTextColor,
-                    ),
-                  ],
-                ),
-                buildDivider(),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Row(
-                      children: [
-                        Text(
-                          'Tasks: ',
-                          style: ToDoTheme.listSubtitleTextStyle.copyWith(
-                            color: ToDoTheme.calendarColor,
-                          ),
-                        ),
-                        Text(
-                          '${pendingTasks.length}/${completedTasks.length + pendingTasks.length} completed',
-                          style: ToDoTheme.listSubtitleTextStyle.copyWith(
-                            color: ToDoTheme.calendarColor,
-                          ),
-                        )
-                      ],
-                    ),
-                    const Icon(
-                      FlutterIcons.ios_arrow_down_ion,
-                      color: ToDoTheme.primaryTextColor,
-                    ),
-                  ],
-                ),
-                Visibility(
-                  visible: todoController.initialExpand.value,
-                  child: Column(
-                    children: [
-                      ListView(
-                        shrinkWrap: true,
-                        physics: const NeverScrollableScrollPhysics(),
-                        children: pendingTasks,
-                      ),
-                      Row(
-                        children: <Widget>[
-                          SizedBox(
-                            height: 30,
-                            width: 109,
-                            child: ElevatedButton(
-                              onPressed: () {
-                                todoController.toggleExpandBtn();
-                              },
-                              style: ButtonStyle(
-                                foregroundColor:
-                                    MaterialStateProperty.all<Color>(
-                                  ToDoTheme.floatingABColor,
-                                ),
-                                backgroundColor:
-                                    MaterialStateProperty.all<Color>(
-                                  ToDoTheme.secondaryCardColor,
-                                ),
-                                padding: MaterialStateProperty.all<EdgeInsets>(
-                                  const EdgeInsets.only(left: 8),
-                                ),
-                              ),
-                              child: Row(
-                                children: <Widget>[
-                                  Text(
-                                    'Completed (${completedTasks.length})',
-                                    style: ToDoTheme.buttonTextStyle,
-                                    softWrap: false,
-                                  ),
-                                  Expanded(
-                                    child: Icon(
-                                      todoController.expandBtn.value
-                                          ? Icons.expand_more
-                                          : Icons.keyboard_arrow_right,
-                                      size: 14,
-                                      color: ToDoTheme.floatingABColor,
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                          ),
-                          const Spacer(),
-                          TextButton(
-                            style: ButtonStyle(
-                              foregroundColor: MaterialStateProperty.all<Color>(
-                                ToDoTheme.secondaryTextColor,
-                              ),
-                              textStyle: MaterialStateProperty.all<TextStyle>(
-                                ToDoTheme.buttonTextStyle,
-                              ),
-                            ),
-                            onPressed: () {
-                              showDialog(
-                                context: context,
-                                builder: (ctx) {
-                                  return const FractionallySizedBox(
-                                    child: Material(
-                                      type: MaterialType.transparency,
-                                      child: AddTaskDialogBox(),
-                                    ),
-                                  );
-                                },
-                              );
-                            },
-                            child: const Text('+ Add Task'),
-                          ),
-                        ],
-                      ),
-                      if (todoController.expandBtn.value)
-                        Padding(
-                          padding: const EdgeInsets.symmetric(vertical: 10),
-                          child: ListView(
-                            shrinkWrap: true,
-                            physics: const NeverScrollableScrollPhysics(),
-                            children: completedTasks,
-                          ),
-                        ),
-                    ],
-                  ),
-                ),
-              ],
+    return Obx(
+      () => controller.todos.isEmpty
+          ? const Center(
+              heightFactor: 10,
+              child: Text(
+                'You do not have any todos!',
+                style: ToDoTheme.titleTextStyle,
+              ),
+            )
+          : ListView.builder(
+              itemCount: controller.todos.length,
+              shrinkWrap: true,
+              itemBuilder: (BuildContext context, int index) {
+                return GetBuilder<ToDoController>(
+                  id: 'list-item-$index',
+                  builder: (_) {
+                    return TodoCard(
+                      controller: controller,
+                      index: index,
+                      isExpanded: controller.cardExpand,
+                      expandBtn: controller.expandBtn,
+                    );
+                  },
+                );
+              },
             ),
-          ),
+    );
+  }
+}
+
+class TodoCard extends StatelessWidget {
+  const TodoCard({
+    super.key,
+    required this.controller,
+    required this.index,
+    required this.isExpanded,
+    required this.expandBtn,
+  });
+  final ToDoController controller;
+  final int index;
+  final bool isExpanded;
+  final bool expandBtn;
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      elevation: 0,
+      margin: const EdgeInsets.all(8),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(15.0),
+      ),
+      color: ToDoTheme.secondaryColor,
+      child: Padding(
+        padding: const EdgeInsets.symmetric(
+          vertical: 8.0,
+          horizontal: 15.0,
+        ),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            _HeaderWidget(
+              title: controller.todos[index].name,
+              team: controller.todos[index].team,
+            ),
+            _DescriptionWidget(
+              description: controller.todos[index].description,
+            ),
+            const _DividerWidget(),
+            _CommentsWidget(context: context),
+            const _DividerWidget(),
+            _TasksRatioWidget(
+              controller: controller,
+              index: index,
+              isExpanded: isExpanded,
+              todo: controller.todos[index],
+            ),
+            Obx(
+              () => _TasksWidget(
+                controller: controller,
+                index: index,
+                context: context,
+                todo: controller.todos[index],
+                isExpanded: isExpanded,
+                expandBtn: expandBtn,
+              ),
+            )
+          ],
         ),
       ),
     );
   }
+}
 
-  Widget buildSubtitle() {
-    return SizedBox(
-      child: widget.subtitle.length > 80
-          ? ExpandableText(widget.subtitle)
-          : Text(
-              widget.subtitle,
-              style: ToDoTheme.listSubtitleTextStyle.copyWith(
-                color: ToDoTheme.calendarColor,
+class _TasksWidget extends StatelessWidget {
+  const _TasksWidget({
+    required this.controller,
+    required this.index,
+    required this.context,
+    required this.todo,
+    required this.isExpanded,
+    required this.expandBtn,
+  });
+  final ToDoController controller;
+  final int index;
+  final BuildContext context;
+  final ToDoList todo;
+  final bool isExpanded;
+  final bool expandBtn;
+
+  @override
+  Widget build(BuildContext context) {
+    List<ToDoTask> _pendingTasks =
+        todo.tasks.where((element) => element.progressPercent < 100).toList();
+    List<ToDoTask> _completedTasks =
+        todo.tasks.where((element) => element.progressPercent >= 100).toList();
+    return Visibility(
+      visible: isExpanded,
+      child: Column(
+        children: <Widget>[
+          ListView(
+            shrinkWrap: true,
+            physics: const NeverScrollableScrollPhysics(),
+            children: List.generate(
+              _pendingTasks.length,
+              (index) => ToDoTaskView(
+                task: _pendingTasks[index],
+                todoList: todo,
               ),
             ),
+          ),
+          Row(
+            children: <Widget>[
+              ElevatedButton(
+                onPressed: () => controller.toggleExpandBtn(index, expandBtn),
+                style: ButtonStyle(
+                  backgroundColor: MaterialStateProperty.all<Color>(
+                    ToDoTheme.secondaryCardColor,
+                  ),
+                  shape: MaterialStateProperty.all<OutlinedBorder>(
+                    RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(5),
+                    ),
+                  ),
+                ),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: <Widget>[
+                    Text(
+                      'Completed (${controller.getCompletedTasks(todo)})',
+                      style: ToDoTheme.buttonTextStyle.copyWith(
+                        color: ToDoTheme.floatingABColor,
+                      ),
+                      softWrap: false,
+                    ),
+                    Icon(
+                      expandBtn
+                          ? Icons.expand_more
+                          : Icons.keyboard_arrow_right,
+                      size: 14,
+                      color: ToDoTheme.floatingABColor,
+                    ),
+                  ],
+                ),
+              ),
+              const Spacer(),
+              TextButton(
+                style: ButtonStyle(
+                  foregroundColor: MaterialStateProperty.all<Color>(
+                    ToDoTheme.secondaryTextColor,
+                  ),
+                  textStyle: MaterialStateProperty.all<TextStyle>(
+                    ToDoTheme.buttonTextStyle,
+                  ),
+                ),
+                onPressed: () {
+                  showDialog(
+                    context: context,
+                    builder: (ctx) {
+                      return FractionallySizedBox(
+                        child: Material(
+                          type: MaterialType.transparency,
+                          child: AddTaskDialogBox(
+                            toDoList: todo,
+                          ),
+                        ),
+                      );
+                    },
+                  );
+                },
+                child: const Text('+ Add Task'),
+              ),
+            ],
+          ),
+          Visibility(
+            visible: expandBtn,
+            child: Padding(
+              padding: const EdgeInsets.symmetric(vertical: 10),
+              child: ListView(
+                shrinkWrap: true,
+                physics: const NeverScrollableScrollPhysics(),
+                children: List.generate(
+                  _completedTasks.length,
+                  (index) => ToDoTaskView(
+                    task: _completedTasks[index],
+                    todoList: todo,
+                  ),
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
     );
   }
+}
 
-  Widget buildDivider() {
+class _TasksRatioWidget extends StatelessWidget {
+  const _TasksRatioWidget({
+    required this.controller,
+    required this.index,
+    required this.isExpanded,
+    required this.todo,
+  });
+  final ToDoController controller;
+  final int index;
+  final bool isExpanded;
+  final ToDoList todo;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: <Widget>[
+        const Text(
+          'Task: ',
+          style: ToDoTheme.listSubtitleTextStyle,
+        ),
+        Text(
+          '${controller.getCompletedTasks(todo)}/${todo.tasks.length} completed',
+          style: ToDoTheme.listSubtitleTextStyle.copyWith(
+            color: ToDoTheme.calendarColor,
+          ),
+        ),
+        const Spacer(),
+        IconButton(
+          onPressed: () => controller.toggleCardExpand(index, isExpanded),
+          icon: Icon(
+            isExpanded
+                ? FlutterIcons.ios_arrow_up_ion
+                : FlutterIcons.ios_arrow_down_ion,
+          ),
+          color: ToDoTheme.primaryTextColor,
+        ),
+      ],
+    );
+  }
+}
+
+class _CommentsWidget extends StatelessWidget {
+  const _CommentsWidget({
+    required this.context,
+  });
+
+  final BuildContext context;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.start,
+      children: [
+        const Icon(
+          FlutterIcons.heart_evi,
+          color: ToDoTheme.primaryTextColor,
+        ),
+        Padding(
+          padding: const EdgeInsets.only(left: 8, top: 3),
+          child: InkWell(
+            onTap: () {},
+            child: SvgPicture.asset(
+              'assets/images/message.svg',
+              color: Colors.white,
+              height: 18,
+              width: 18,
+            ),
+          ),
+        ),
+        const Spacer(),
+        const Icon(
+          FlutterIcons.bookmark_border_mdi,
+          color: ToDoTheme.primaryTextColor,
+        ),
+      ],
+    );
+  }
+}
+
+class _DescriptionWidget extends StatelessWidget {
+  const _DescriptionWidget({
+    required this.description,
+  });
+
+  final String? description;
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      child: (description != null || description!.isNotEmpty)
+          ? description!.length > 80
+              ? ExpandableText(description!)
+              : Text(description!, style: ToDoTheme.descriptionTextStyle)
+          : const SizedBox.shrink(),
+    );
+  }
+}
+
+class _DividerWidget extends StatelessWidget {
+  const _DividerWidget();
+
+  @override
+  Widget build(BuildContext context) {
     return const Divider(
       color: ToDoTheme.listDividerColor,
       indent: 0,
       endIndent: 0,
       thickness: 1,
+    );
+  }
+}
+
+class _HeaderWidget extends StatelessWidget {
+  const _HeaderWidget({
+    required this.title,
+    required this.team,
+  });
+
+  final String title;
+  final Team? team;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 5),
+      child: Row(
+        children: [
+          Text(
+            title,
+            style: ToDoTheme.listTitleTextStyle,
+          ),
+          const SizedBox(
+            width: 8.0,
+          ),
+          team != null
+              ? Container(
+                  padding: const EdgeInsets.all(8.0),
+                  decoration: const BoxDecoration(
+                    color: AppCommonTheme.secondaryColor,
+                    borderRadius: BorderRadius.all(Radius.circular(10.0)),
+                  ),
+                  child: Text(
+                    team!.name!,
+                    style: ToDoTheme.listTagTextStyle,
+                  ),
+                )
+              : const SizedBox.shrink(),
+        ],
+      ),
     );
   }
 }
